+++ x/gamm/types/key.go
@@ -25,6 +25,10 @@ var (
 	KeyTotalLiquidity = []byte{0x03}
 )
 
+func GetDenomPrefix(denom string) []byte {
+	return append(KeyTotalLiquidity, []byte(denom)...)
+}
+
 func GetPoolShareDenom(poolId uint64) string {
 	return fmt.Sprintf("gamm/pool/%d", poolId)
 }