+++ x/incentives/client/cli/cli_test.go
@@ -206,6 +206,40 @@ func (s *IntegrationTestSuite) TestGetCmdActiveGauges() {
 	}
 }
 
+func (s *IntegrationTestSuite) TestGetCmdActiveGaugesPerDenom() {
+	val := s.network.Validators[0]
+
+	testCases := []struct {
+		name      string
+		expectErr bool
+		respType  proto.Message
+	}{
+		{
+			"query active gauges per denom",
+			false, &types.ActiveGaugesPerDenomResponse{},
+		},
+	}
+
+	for _, tc := range testCases {
+		tc := tc
+
+		s.Run(tc.name, func() {
+			cmd := cli.GetCmdActiveGaugesPerDenom()
+			clientCtx := val.ClientCtx
+
+			args := []string{s.cfg.BondDenom}
+
+			out, err := clitestutil.ExecTestCLICmd(clientCtx, cmd, args)
+			if tc.expectErr {
+				s.Require().Error(err)
+			} else {
+				s.Require().NoError(err, out.String())
+				s.Require().NoError(clientCtx.JSONMarshaler.UnmarshalJSON(out.Bytes(), tc.respType), out.String())
+			}
+		})
+	}
+}
+
 func (s *IntegrationTestSuite) TestGetCmdUpcomingGauges() {
 	val := s.network.Validators[0]